backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic>=1.13.0

# Validation & Settings
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Redis & Caching
redis>=5.0.0
aioredis>=2.0.0

# Background Jobs
celery>=5.3.0

# AI/LLM
openai>=2.0.0
pgvector>=0.3.0

# HTTP Client
httpx>=0.27.0

# Logging & Monitoring
python-json-logger>=2.0.0
prometheus-client>=0.19.0

# Optimization & PDF
reportlab==4.2.2
ortools==9.10.4067

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
PyJWT==2.10.1
httpx>=0.27.0
python-multipart>=0.0.9
